pyproject.toml
[project]
name = "cs178-goop-squad-cifar10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.6.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
]
